refactor(moyopy): move subgroup bindings
