+++ moyopy/src/subgroup.rs
@@ -1,0 +1,6 @@
+mod translationengleiche_subgroup;
+
+pub use translationengleiche_subgroup::{
+    PyTranslationengleicheSubgroup, PyTranslationengleicheSubgroupConjugacyClass,
+    PyTranslationengleicheSubgroupConjugate, enumerate_translationengleiche_subgroups,
+};
